% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_selina_model.R
\name{load_selina_model}
\alias{load_selina_model}
\title{load_selina_model}
\usage{
load_selina_model(path_model)
}
\arguments{
\item{path_model}{Path for loading SELINA model. Please put '.pth' and 'rds' in the same folder.}
}
\value{
A trained MADA model.
}
\description{
load_selina_model
}
