% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{query_predict}
\alias{query_predict}
\title{Fine-tuning and predict for the query data}
\usage{
query_predict(query_expr, model, path_out, outprefix, disease, mode)
}
\arguments{
\item{query_expr}{File path of the query data matrix}

\item{model}{File path of the pre-trained model}

\item{path_out}{File path of the output files.}

\item{outprefix}{Prefix of the output files. DEFAULT: query.}

\item{disease}{Depend on your data is in some disease condition or not, choose from 'TRUE' or 'FALSE'. DEFAULT: 'FALSE'.}

\item{mode}{Single-cell level input or cluster level input, choose from 'single' or 'cluster'. DEFAULT: 'single'.}
}
\value{
Prediction results and corresponding probability for each query cell.
}
\description{
Fine-tuning and predict for the query data
}
