% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model.R
\name{read_model}
\alias{read_model}
\title{Read in model trained by yourself}
\usage{
read_model(path_model)
}
\arguments{
\item{path_model}{Path for model file suffixed with '.pt'.}
}
\value{
A trained MADA model.
}
\description{
Read in model trained by yourself
}
