% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_model}
\alias{save_model}
\title{save_model}
\usage{
save_model(model, path_out, prefix = "train")
}
\arguments{
\item{model}{Output of train_model}

\item{path_out}{Directory for saving model.}

\item{prefix}{Prefix of the output files. DEFAULT: train}
}
\value{
None
}
\description{
save_model
}
\examples{
save_model(model, path_out, prefix)
}
