% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_model}
\alias{train_model}
\title{Train model based on your own data}
\usage{
train_model(seuratlist, path_in, path_out, prefix = "pre-trained")
}
\arguments{
\item{seuratlist}{A list of seurat object.}

\item{path_in}{File path of reference meta files.}

\item{path_out}{File path of the output model.}

\item{prefix}{Prefix of the output files. DEFAULT: pre-trained.}
}
\value{
An MADA model.
}
\description{
Train model based on your own data
}
