% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_selina_model.R
\name{load_selina_model}
\alias{load_selina_model}
\title{Load model trained by SELINA}
\usage{
load_selina_model(tissue)
}
\arguments{
\item{tissue}{Choose which tissue model you want to load, eg: Pancreas. Please see documentation for tissue name.}
}
\value{
A trained MADA model.
}
\description{
Load model trained by SELINA
}
