% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{query_predict}
\alias{query_predict}
\title{Fine-tuning and predict for the query data}
\usage{
query_predict(
  queryObj,
  model,
  path_out,
  outprefix,
  disease = FALSE,
  cell_cutoff = 5,
  prob_cutoff = 0.9
)
}
\arguments{
\item{queryObj}{Seurat object for query data.}

\item{model}{File path of the pre-trained model.}

\item{path_out}{File path of the output files.}

\item{outprefix}{Prefix of the output files. DEFAULT: query.}

\item{disease}{Depend on your data is in some disease condition or not, choose from 'TRUE' or 'FALSE'. DEFAULT: 'FALSE'.}

\item{cell_cutoff}{Cutoff for number of cells with the same cell type in 10 nearest neighbor cells(only used when the input is single-cell level expression matrix). DEFAULT: 5.}

\item{prob_cutoff}{Cutoff for prediction probability. DEFAULT: 0.9.}
}
\value{
A list which includes prediction results and corresponding probability for query data. This step will generate eight files.
}
\description{
Fine-tuning and predict for the query data
}
\examples{
query_result <- query_predict(queryObj,
                              model,
                              path_out,
                              outprefix = 'demo', 
                              disease = FALSE, 
                              cell_cutoff = 5,
                              prob_cutoff = 0.9)
}
