% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{query_predict}
\alias{query_predict}
\title{Fine-tuning and predict for the query data}
\usage{
query_predict(
  seuratobj,
  model,
  path_meta,
  path_out,
  outprefix,
  disease = FALSE,
  mode = "single",
  seurat,
  cell_cutoff = 5,
  prob_cutoff = 0.9
)
}
\arguments{
\item{model}{File path of the pre-trained model.}

\item{path_out}{File path of the output files.}

\item{outprefix}{Prefix of the output files. DEFAULT: query.}

\item{disease}{Depend on your data is in some disease condition or not, choose from 'TRUE' or 'FALSE'. DEFAULT: 'FALSE'.}

\item{mode}{Single-cell level input or cluster level input, choose from 'single' or 'cluster'. DEFAULT: 'single'.}

\item{seurat}{File path of the seurat object.}

\item{cell_cutoff}{Cutoff for number of cells with the same cell type in 10 nearest neighbor cells(only used when the input is single-cell level expression matrix). DEFAULT: 5.}

\item{prob_cutoff}{Cutoff for prediction probability. DEFAULT: 0.9.}

\item{query_expr}{File path of the query data matrix(rds format).}
}
\value{
Prediction results and corresponding probability for query. This step will output eight files for 'single' mode and four files for 'cluster' mode
}
\description{
Fine-tuning and predict for the query data
}
