% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model.R
\name{read_model}
\alias{read_model}
\title{read_model}
\usage{
read_model(path_model, path_meta)
}
\arguments{
\item{path_model}{Path for loading model.}

\item{path_meta}{Path for read meta file (rds format)}
}
\value{
A MADA object.
}
\description{
read_model
}
\examples{
model <- read_model(path_model)
}
