% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_model}
\alias{train_model}
\title{Train model based on your own data}
\usage{
train_model(path_in, disease)
}
\arguments{
\item{path_in}{Folder includes expression and meta files.}

\item{disease}{Depend on your data is in some disease condition or not, choose from 'TRUE' or 'FALSE'. DEFAULT: 'FALSE'.}
}
\value{
An MADA model.
}
\description{
Train model based on your own data
}
