% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_model}
\alias{train_model}
\title{Train model based on your own data}
\usage{
train_model(seuratlist)
}
\arguments{
\item{seuratlist}{A list of seurat object for training.}
}
\value{
An MADA model.
}
\description{
Train model based on your own data
}
